// src/services/mock-api.ts
export interface Task {
  id: string;
  title: string;
  description?: string;
  isCompleted: boolean;
  dueDate?: Date; // ISO 8601 Date String
}